% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vincentverse-package.R
\docType{package}
\name{vincentverse-package}
\alias{vincentverse}
\alias{vincentverse-package}
\title{vincentverse: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JozefRivest/vincentverse}
  \item Report bugs at \url{https://github.com/JozefRivest/vincentverse/issues}
}

}
\author{
\strong{Maintainer}: Jozef Rivest \email{jrivest@umich.edu} (\href{https://orcid.org/0000-0002-3195-3838}{ORCID}) [copyright holder]

}
\keyword{internal}
