% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{vincentverse_deps}
\alias{vincentverse_deps}
\title{List all vincentverse dependencies}
\usage{
vincentverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
vincentverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all vincentverse dependencies
}
