% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vincentverse_packages}
\alias{vincentverse_packages}
\title{List all packages in the vincentverse}
\usage{
vincentverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include vincentverse in the list?}
}
\description{
List all packages in the vincentverse
}
\examples{
vincentverse_packages()
}
